feat: bootstrap pr body validation gate
