+++ .github/workflows/validate.yml
@@ -27,6 +27,6 @@ jobs:
         if: github.event_name == 'pull_request'
         env:
           PR_BODY: ${{ github.event.pull_request.body || '' }}
-        run: python scripts/validate_pr_body.py --body "$PR_BODY"
+        run: python skills/ora-et-labora/scripts/validate_pr_body.py --body "$PR_BODY"
       - name: Validate suite
         run: python scripts/validate_all.py
