+++ .project/logs/10.md
@@ -1,0 +1,5 @@
+# Task Log: 10
+
+- 2026-04-24 | state:init | branch:feat/10-bootstrap-pr-body-validation | issue:#10 | note:Workspace initialized.
+- 2026-04-24 | verify:local-pass | command:`python scripts/validate_all.py` | note:21 tests passed and all skills validated after moving the PR validator into the shipped skill payload.
+- 2026-04-24 | verify:bootstrap-smoke-pass | command:`python scripts/validate_pr_body.py --body-file <bootstrapped-body>` | note:Bootstrapped repos receive a runnable validator plus standalone PR-body workflow asset.
