+++ .project/todo/10/00_brainstorm.md
@@ -1,0 +1,41 @@
+# Propagate PR body validation through bootstrap assets: Brainstorm
+
+- Date: 2026-04-24
+- Issue: [#10](https://github.com/emmepra/ora-et-labora/issues/10)
+- Kind: feature
+- Module: 10
+
+## Problem
+
+- 
+
+## Desired Outcome
+
+- 
+
+## Constraints
+
+- 
+
+## Blueprint Fit Check
+
+- Relevant blueprint docs:
+- Feasibility:
+- Conflicts or missing decisions:
+
+## Options Considered
+
+- Option A:
+- Option B:
+
+## Chosen Direction
+
+- 
+
+## Risks / Unknowns
+
+- 
+
+## Acceptance Checks
+
+-
