+++ .project/todo/10/CURRENT.md
@@ -1,0 +1,14 @@
+# Current State
+
+- Issue: [#10](https://github.com/emmepra/ora-et-labora/issues/10)
+- Title: Propagate PR body validation through bootstrap assets
+- Kind: feature
+- Module: 10
+- Branch: feat/10-bootstrap-pr-body-validation
+- Status: implementation complete; ready to open PR
+- PR: pending
+- Last verification: `python scripts/validate_all.py` passed; skill-owned validator and bootstrapped-copy smoke checks passed
+- Browser evidence: not applicable
+- Next step: commit, push, and open the PR to `dev`; then enable auto-merge once required checks are pending/passing
+- Blockers: none
+- Files touched: `.github/workflows/validate.yml`, `scripts/validate_pr_body.py`, `skills/ora-et-labora/SKILL.md`, `skills/ora-et-labora/scripts/bootstrap_repo_templates.py`, `skills/ora-et-labora/scripts/validate_pr_body.py`, `skills/ora-et-labora/assets/bootstrap/.github/workflows/validate-pr-body.yml`, `skills/repo-bootstrap/SKILL.md`, `tests/test_bootstrap_repo_templates.py`, `tests/test_template_enforcement_policy.py`, `tests/test_validate_pr_body.py`
