+++ tests/test_bootstrap_repo_templates.py
@@ -26,16 +26,69 @@ class BootstrapRepoTemplatesTests(unittest.TestCase):
 
             self.assertTrue((repo_root / ".github" / "PULL_REQUEST_TEMPLATE.md").exists())
             self.assertTrue((repo_root / ".github" / "ISSUE_TEMPLATE" / "config.yml").exists())
+            self.assertTrue((repo_root / ".github" / "workflows" / "validate-pr-body.yml").exists())
             self.assertTrue((repo_root / ".project" / "blueprint" / "00_workflow.md").exists())
+            self.assertTrue((repo_root / "scripts" / "validate_pr_body.py").exists())
             pr_template = (repo_root / ".github" / "PULL_REQUEST_TEMPLATE.md").read_text()
             issue_config = (repo_root / ".github" / "ISSUE_TEMPLATE" / "config.yml").read_text()
+            pr_workflow = (repo_root / ".github" / "workflows" / "validate-pr-body.yml").read_text()
             self.assertIn("## Linked Issue", pr_template)
             self.assertIn("Closes #", pr_template)
             self.assertIn("blank_issues_enabled: false", issue_config)
+            self.assertIn("python scripts/validate_pr_body.py", pr_workflow)
             gitignore = (repo_root / ".gitignore").read_text()
             self.assertIn("Visibility profile: private", gitignore)
             self.assertIn(".project/worktrees/", gitignore)
             self.assertNotIn("\n.project/\n", f"\n{gitignore}")
+
+            body = repo_root / "body.md"
+            body.write_text(
+                """## Summary
+
+- Bootstrapped smoke test.
+
+## Why
+
+Ensure the copied validator runs in a target repo.
+
+## Linked Issue
+
+Closes #1
+
+## Verification
+
+- Local: smoke
+- Browser: not applicable
+- Browser evidence: not applicable
+- CI: pending
+
+## Auto-Merge Eligibility
+
+- Not applicable.
+
+## Blueprint Updates
+
+None.
+
+## Risks / Rollback
+
+Revert the bootstrap if needed.
+
+## Follow-ups
+
+- None.
+"""
+            )
+            subprocess.run(
+                [
+                    "python",
+                    str(repo_root / "scripts" / "validate_pr_body.py"),
+                    "--body-file",
+                    str(body),
+                ],
+                check=True,
+                cwd=repo_root,
+            )
 
     def test_public_visibility_keeps_project_state_local(self) -> None:
         with tempfile.TemporaryDirectory() as tmp:
