+++ tests/test_template_enforcement_policy.py
@@ -34,7 +34,11 @@ class TemplateEnforcementPolicyTests(unittest.TestCase):
     def test_validate_workflow_runs_pr_body_gate(self) -> None:
         workflow = (REPO_ROOT / ".github" / "workflows" / "validate.yml").read_text()
         self.assertIn("Validate PR body", workflow)
-        self.assertIn("scripts/validate_pr_body.py", workflow)
+        self.assertIn("skills/ora-et-labora/scripts/validate_pr_body.py", workflow)
+
+    def test_skill_contains_pr_body_validator(self) -> None:
+        validator = (REPO_ROOT / "skills" / "ora-et-labora" / "scripts" / "validate_pr_body.py")
+        self.assertTrue(validator.exists())
 
 
 if __name__ == "__main__":
